move all tests into a tests folder
