--- src/components/LoaderInline/LoaderInline.cy.tsx
@@ -1,8 +1,0 @@
-import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
-import { LoaderInline } from './LoaderInline';
-
-describe('Inline loader', () => {
-  beforeEach(() => {
-    CustomMount(<LoaderInline></LoaderInline>);
-  });
-});
