--- src/components/LoaderOverlay/LoaderOverlay.cy.tsx
@@ -1,8 +1,0 @@
-import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
-import { LoaderOverlay } from './LoaderOverlay';
-
-describe('Loader Overlay', () => {
-  beforeEach(() => {
-    CustomMount(<LoaderOverlay isEnabled></LoaderOverlay>);
-  });
-});
