--- src/components/ToastWrapper/ToastWrapper.cy.tsx
@@ -1,8 +1,0 @@
-import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
-import { ToastWrapper } from './ToastWrapper';
-
-describe('Toast Wrapper', () => {
-  beforeEach(() => {
-    CustomMount(<ToastWrapper></ToastWrapper>);
-  });
-});
