--- src/pages/Checkout/Checkout.cy.tsx
@@ -1,8 +1,0 @@
-import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
-import { CheckoutPage } from './Checkout';
-
-describe('Checkout Page Tests', () => {
-  it('mounts', () => {
-    CustomMount(<CheckoutPage />);
-  });
-});
