--- src/pages/Home/Home.cy.tsx
@@ -1,16 +1,0 @@
-import { RouteConstants } from 'utilities/route-constants';
-import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
-import { HomePage } from './Home';
-
-describe('HomePage tests', () => {
-  it('mounts', () => {
-    CustomMount(<HomePage />);
-  });
-
-  it('Navigates to menu when "Menu" button is clicked', () => {
-    CustomMount(<HomePage />);
-
-    cy.get('button').click();
-    cy.location('pathname').should('contain', RouteConstants.MENU);
-  });
-});
