--- src/pages/Menu/Menu.cy.tsx
@@ -1,8 +1,0 @@
-import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
-import { MenuPage } from './Menu';
-
-describe('Menu Page Tests', () => {
-  it('mounts', () => {
-    CustomMount(<MenuPage />);
-  });
-});
