+++ src/tests/components/Container/Container.cy.tsx
@@ -1,0 +1,12 @@
+import { Container } from 'components/Container/Container';
+import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
+
+describe('container', () => {
+  beforeEach(() => {
+    CustomMount(<Container>Container</Container>);
+  });
+
+  it('displays text', () => {
+    cy.getBySel('container').should('contain', 'Container');
+  });
+});
