+++ src/tests/components/LoaderInline/LoaderInline.cy.tsx
@@ -1,0 +1,8 @@
+import { LoaderInline } from 'components/LoaderInline/LoaderInline';
+import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
+
+describe('Inline loader', () => {
+  beforeEach(() => {
+    CustomMount(<LoaderInline></LoaderInline>);
+  });
+});
