+++ src/tests/components/LoaderOverlay/LoaderOverlay.cy.tsx
@@ -1,0 +1,8 @@
+import { LoaderOverlay } from 'components/LoaderOverlay/LoaderOverlay';
+import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
+
+describe('Loader Overlay', () => {
+  beforeEach(() => {
+    CustomMount(<LoaderOverlay isEnabled></LoaderOverlay>);
+  });
+});
