+++ src/tests/components/ToastWrapper/ToastWrapper.cy.tsx
@@ -1,0 +1,8 @@
+import { ToastWrapper } from 'components/ToastWrapper/ToastWrapper';
+import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
+
+describe('Toast Wrapper', () => {
+  beforeEach(() => {
+    CustomMount(<ToastWrapper></ToastWrapper>);
+  });
+});
