+++ src/tests/pages/Checkout/Checkout.cy.tsx
@@ -1,0 +1,8 @@
+import { CheckoutPage } from 'pages/Checkout/Checkout';
+import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
+
+describe('Checkout Page Tests', () => {
+  it('mounts', () => {
+    CustomMount(<CheckoutPage />);
+  });
+});
