+++ src/tests/pages/Menu/Menu.cy.tsx
@@ -1,0 +1,8 @@
+import { MenuPage } from 'pages/Menu/Menu';
+import CustomMount from '../../testHelpers/cypressHelpers/CustomMount';
+
+describe('Menu Page Tests', () => {
+  it('mounts', () => {
+    CustomMount(<MenuPage />);
+  });
+});
